move sha1 to unsafe crate
